Updating PXOnteTapViewController
